# ext/bcmath: Avoid unnecessary `memset`