--- a/ext/bcmath/libbcmath/src/doaddsub.c
+++ b/ext/bcmath/libbcmath/src/doaddsub.c
@@ -121,7 +121,7 @@ bc_num _bc_do_add(bc_num n1, bc_num n2, size_t scale_min)
    returned.  The signs of N1 and N2 are ignored.  Also, N1 is
    assumed to be larger than N2.  SCALE_MIN is the minimum scale
    of the result. */
-bc_num _bc_do_sub(bc_num n1, bc_num n2, size_t scale_min)
+bc_num _bc_do_sub(bc_num n1, bc_num n2)
 {
 	bc_num diff;
 	/* The caller is guaranteed that n1 is always large. */
@@ -137,7 +137,7 @@ bc_num _bc_do_sub(bc_num n1, bc_num n2, size_t scale_min)
 	char *n1ptr, *n2ptr, *diffptr;
 
 	/* Allocate temporary storage. */
-	diff = bc_new_num (diff_len, MAX(diff_scale, scale_min));
+	diff = bc_new_num_nonzeroed(diff_len, diff_scale);
 
 	/* Initialize the subtract. */
 	n1ptr = (char *) (n1->n_value + n1->n_len + n1->n_scale - 1);