@nielsdos 2024-05-08
Intuitively, this speed gain feels strange because memset only takes up 0.35% of runtime on my machine for 1.php for example. Furthermore, I actually see a slowdown when applying this patch on current master:

```
Benchmark 1: ./sapi/cli/php 1.php
  Time (mean ± σ):     436.2 ms ±   6.2 ms    [User: 432.7 ms, System: 2.8 ms]
  Range (min … max):   430.3 ms … 450.7 ms    10 runs
 
Benchmark 2: ./sapi/cli/php_old 1.php
  Time (mean ± σ):     418.8 ms ±   8.3 ms    [User: 415.5 ms, System: 2.5 ms]
  Range (min … max):   412.3 ms … 440.9 ms    10 runs
 
Summary
  ./sapi/cli/php_old 1.php ran
    1.04 ± 0.03 times faster than ./sapi/cli/php 1.php


Benchmark 1: ./sapi/cli/php 2.php
  Time (mean ± σ):     499.6 ms ±   3.1 ms    [User: 496.0 ms, System: 2.7 ms]
  Range (min … max):   494.7 ms … 503.3 ms    10 runs
 
Benchmark 2: ./sapi/cli/php_old 2.php
  Time (mean ± σ):     475.8 ms ±   3.4 ms    [User: 473.5 ms, System: 1.4 ms]
  Range (min … max):   472.0 ms … 480.1 ms    10 runs
 
Summary
  ./sapi/cli/php_old 2.php ran
    1.05 ± 0.01 times faster than ./sapi/cli/php 2.php

```

I see a similar slowdown for your old benchmark script with microtime.
It doesn't make sense that the execution becomes slower, I think something deeper is going on especially as you see the opposite effect...