@SakiTakamachi 2024-05-08
Thanks. It seems the measurements on my machine are unreliable....

I will try to prepare a stable measurement environment such as EC2.