@SakiTakamachi 2024-05-08
@nielsdos 
I merged the latest master and then compared master and this branch on EC2.

master:
```
hyperfine "php 1.php" --warmup 10
Time (mean ± σ):     654.7 ms ±   3.0 ms    [User: 650.5 ms, System: 2.6 ms]
Range (min … max):   650.1 ms … 659.2 ms    10 runs

hyperfine "php 2.php" --warmup 10
Time (mean ± σ):     769.4 ms ±   5.4 ms    [User: 765.8 ms, System: 2.0 ms]
Range (min … max):   762.6 ms … 781.8 ms    10 runs

hyperfine "php 3.php" --warmup 10
Time (mean ± σ):     910.3 ms ±  13.3 ms    [User: 905.6 ms, System: 2.8 ms]
Range (min … max):   896.8 ms … 934.6 ms    10 runs

php old.php // my old bench
1.6298861503601
1.9048039913177
2.2188358306885
```

this branch:
```
hyperfine "php 1.php" --warmup 10
Time (mean ± σ):     605.0 ms ±   2.5 ms    [User: 601.5 ms, System: 2.0 ms]
Range (min … max):   601.3 ms … 609.0 ms    10 runs

hyperfine "php 2.php" --warmup 10
Time (mean ± σ):     717.5 ms ±   5.0 ms    [User: 713.1 ms, System: 2.7 ms]
Range (min … max):   712.5 ms … 730.7 ms    10 runs

hyperfine "php 3.php" --warmup 10
Time (mean ± σ):     863.4 ms ±   9.3 ms    [User: 858.7 ms, System: 2.6 ms]
Range (min … max):   851.4 ms … 878.6 ms    10 runs

php old.php // my old bench
1.5056660175323
1.7816500663757
2.116672039032
```