@SakiTakamachi 2024-05-09
Is the compiler gcc?

I am compiling both with gcc on ubuntu. It might be worth checking if a different assembly is being generated.