@SakiTakamachi 2024-05-09
My machine is 9, but the EC2 is 13.

Could you please send me the assembler code later when you have time? I'd like to compare it with mine.